When first visiting region page, always start the table out to be filtered to the auto-suggested best gene
